Revert "Revert "feat: CellGuide marker gene api refinements" (#7009)"
This reverts commit 9b0cafa0213c1f1ae283b7773bebe60047de2491.
