+++ backend/cellguide/api/cellguide-api.yml
@@ -1,13 +1,13 @@
 openapi: 3.0.0
 info:
   version: "1.0.0"
-  title: Chan Zuckerberg Initiative CELLxGENE Discover API
+  title: Chan Zuckerberg Initiative CELLxGENE CellGuide API
   termsOfService: https://cellxgene.cziscience.com/tos
   contact:
     name: API Support
     email: cellxgene@chanzuckerberg.com
   description: |
-    ## Uploads for CellGuide Descriptions
+    ## Overview of Available Endpoints for the CELLxGENE CellGuide API
 servers:
   - description: Production environment
     url: https://api.cellxgene.cziscience.com/
@@ -49,29 +49,11 @@ paths:
           $ref: "#/components/responses/403"
   /v1/marker_genes:
     get:
-      summary: Get marker genes for a given organism, tissue (optional), and cell type (optional)
+      summary: |
+        Get all CellGuide marker genes.
       operationId: backend.cellguide.api.v1.marker_genes.actions.get
       tags:
         - CellGuide
-      parameters:
-        - in: query
-          name: organism
-          required: true
-          schema:
-            $ref: "#/components/schemas/organism_ontology_term_id_or_label"
-          description: The ontology ID or name of the organism.
-        - in: query
-          name: tissue
-          required: false
-          schema:
-            $ref: "#/components/schemas/tissue_ontology_term_id_or_label"
-          description: The ontology ID or name of the tissue, optional. Use 'All Tissues' to get tissue-agnostic marker genes.
-        - in: query
-          name: cell_type
-          required: false
-          schema:
-            $ref: "#/components/schemas/cell_ontology_term_id"
-          description: The ontology ID or name of the cell type, optional.
       responses:
         "200":
           description: "Successful retrieval of marker genes data."
@@ -88,40 +70,21 @@ paths:
 
 components:
   schemas:
-    organism_ontology_term_id_or_label:
-      description: Organism ontology term ID or label
-      example: Homo sapiens
-      type: string
-    tissue_ontology_term_id_or_label:
-      description: Tissue ontology term ID or label
-      example: All Tissues
-      type: string
-    cell_ontology_term_id:
-      description: Cell ontology term ID
-      example: CL:0000540
-      type: string
     cell_ontology_term_id_fs:
       description: File system-compatible cell ontology term ID (with underscores)
       example: CL_0000030
       type: string
     marker_gene_data:
       description: Details of a marker gene, which can vary in structure based on the specificity of the query (organism, tissue, cell type). Can be a nested structure or a direct list.
-      oneOf:
-        - type: object
-          additionalProperties: # This allows any property name for tissue
-            type: object
-            additionalProperties: # This allows any property name for cell type
-              type: array
-              items:
-                $ref: "#/components/schemas/marker_gene_entry"
-        - type: object
+      type: object
+      additionalProperties: # This allows any property name for organism
+        type: object
+        additionalProperties: # This allows any property name for tissue
+          type: object
           additionalProperties: # This allows any property name for cell type
             type: array
             items:
               $ref: "#/components/schemas/marker_gene_entry"
-        - type: array
-          items:
-            $ref: "#/components/schemas/marker_gene_entry"
     marker_gene_entry:
       type: object
       properties:
